Fix a few more instances of the same warning

+++ tests/c/client_test_put_get_1D.c
@@ -163,7 +163,7 @@ int put_get_1D_tensor_float(size_t* dims, size_t n_dims,
   float* result;
 
   for(size_t i=0; i<dims[0]; i++)
-    tensor[i] = ((float)rand())/RAND_MAX;
+    tensor[i] = ((float)rand())/(float)RAND_MAX;
 
   int r_value = 0;
   r_value = put_get_1D_tensor(client,(void*)tensor,
